backend/requirements.txt
# =============================================================================
# Core Flask dependencies
# =============================================================================
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# =============================================================================
# Database clients
# =============================================================================
supabase==2.10.0
firebase-admin==6.4.0

# =============================================================================
# Security
# =============================================================================
cryptography==42.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

# =============================================================================
# AI Brain dependencies
# =============================================================================
openai>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0

# =============================================================================
# Production WSGI Server with Async I/O
# =============================================================================
gunicorn==21.2.0

# Async workers - choose based on your platform:
# Option 1: gthread (works on all platforms, recommended for Windows)
# No additional package needed - uses Python threads

# Option 2: gevent (Linux/macOS production - uncomment if on Linux)
# gevent>=24.10.3
# greenlet>=3.0.0

# Option 3: eventlet (alternative async - uncomment if preferred)
# eventlet>=0.35.0

# =============================================================================
# Caching (Redis-backed)
# =============================================================================
Flask-Caching==2.3.0
redis==5.0.1

# =============================================================================
# Background Task Processing (Celery)
# =============================================================================
celery[redis]==5.3.6
kombu>=5.3.0

# =============================================================================
# Performance & Compression
# =============================================================================
Flask-Compress==1.14
orjson>=3.9.10  # Fast JSON serialization

# =============================================================================
# Monitoring & Profiling
# =============================================================================
Flask-Profiler==1.8.1
prometheus-flask-exporter==0.23.0

# =============================================================================
# Async Support
# =============================================================================
asgiref==3.7.2
aiohttp>=3.9.1
asyncio-throttle==1.0.2

# =============================================================================
# Load Testing (dev dependency - optional)
# =============================================================================
# locust==2.20.0  # Uncomment for load testing

# =============================================================================
# Retry & Circuit Breaker
# =============================================================================
tenacity==8.2.3
pybreaker==1.0.2

# =============================================================================
# Memory Optimization
# =============================================================================
numpy>=1.24.0  # For vector operations
cachetools==5.3.2  # Advanced caching utilities

# =============================================================================
# Logging & Observability
# =============================================================================
structlog==24.1.0
python-json-logger==2.0.7

# OpenTelemetry (Prometheus metrics for feature gate monitoring)
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-prometheus>=0.48b0
prometheus-client>=0.20.0

# =============================================================================
# Google Sheets Integration (for order sync)
# =============================================================================
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# =============================================================================
# Payment Gateway (Razorpay)
# =============================================================================
razorpay>=1.4.0

# =============================================================================
# PDF Invoice Generation
# =============================================================================
reportlab>=4.0.0

# =============================================================================
# Email OTP (Resend)
# =============================================================================
resend>=2.0.0
